% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imagefilter.R
\name{color_cutoff_filter}
\alias{color_cutoff_filter}
\title{Title Apply a greyscale-based cutoff to convert pixel color}
\usage{
color_cutoff_filter(pixel, cutoff = 128)
}
\arguments{
\item{pixel}{A numeric vector of length 3 representing an RGB pixel.}

\item{cutoff}{A numeric value between 0 and 255. Defaults to 128.}
}
\value{
A numeric vector of length 3 representing the filtered RGB pixel.
}
\description{
Title Apply a greyscale-based cutoff to convert pixel color
}
