% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imagefilter.R
\name{greyscale_filter}
\alias{greyscale_filter}
\title{Title Convert a pixel to greyscale}
\usage{
greyscale_filter(pixel)
}
\arguments{
\item{pixel}{A numeric vector of length 3 representing an RGB pixel.}
}
\value{
A numeric vector of length 3 with identical greyscale values.
}
\description{
Title Convert a pixel to greyscale
}
