% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_cutoff.R
\name{image_cutoff}
\alias{image_cutoff}
\title{image_cutoff}
\usage{
image_cutoff(x, y)
}
\arguments{
\item{x}{image you have}

\item{y}{name of the output figure}
}
\value{
cutoff figures
}
\description{
Takes a rgb image and applies a cut-off filter where pixels with #808080 gets
colored pink #FF00FF and other pixels #000000.
}
\examples{
image_cutoff("https://encrypted-tbn0.gstatic.com/images?q=tbn:ANd9GcRKwyTL9drJIaBbGx_-p8ogSLh_9DcPVDM8mQ&s","test.jpg")

}
