% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_grey.R
\name{image_grey}
\alias{image_grey}
\title{image_grey}
\usage{
image_grey(x, y)
}
\arguments{
\item{x}{image you have}

\item{y}{name of the output figure}
}
\value{
grey figures
}
\description{
Takes a rgb image and converts it to grayscale.
}
\examples{
image_grey("https://jeroen.github.io/images/frink.png","test.jpg")

}
